Move generation classes to correct namespace
