+++ src/NJsonSchema/Generation/DefaultSchemaNameGenerator.cs
@@ -7,10 +7,12 @@
 //-----------------------------------------------------------------------
 
 using System;
+using System.Text.RegularExpressions;
 using Namotion.Reflection;
+using Newtonsoft.Json;
 using NJsonSchema.Annotations;
 
-namespace NJsonSchema
+namespace NJsonSchema.Generation
 {
     /// <summary>The default schema name generator implementation.</summary>
     public class DefaultSchemaNameGenerator : ISchemaNameGenerator
@@ -20,13 +22,19 @@ namespace NJsonSchema
         /// <returns>The new name.</returns>
         public virtual string Generate(Type type)
         {
-            var jsonSchemaAttribute = type.ToCachedType().GetTypeAttribute<JsonSchemaAttribute>();
-            if (!string.IsNullOrEmpty(jsonSchemaAttribute?.Name))
-                return jsonSchemaAttribute.Name;
+            var cachedType = type.ToCachedType();
 
-            //var jsonObjectAttribute = type.GetTypeInfo().GetCustomAttribute<JsonObjectAttribute>();
+            var jsonSchemaAttribute = cachedType.GetTypeAttribute<JsonSchemaAttribute>();
+            if (!string.IsNullOrEmpty(jsonSchemaAttribute?.Name))
+            {
+                return jsonSchemaAttribute.Name;
+            }
+
+            //var jsonObjectAttribute = cachedType.GetTypeAttribute<JsonObjectAttribute>();
             //if (!string.IsNullOrEmpty(jsonObjectAttribute.Title) && Regex.IsMatch(jsonObjectAttribute.Title, "^[a-zA-Z0-9_]*$"))
+            //{
             //    return jsonObjectAttribute.Title;
+            //}
 
             return type.GetDisplayName();
         }
